Move VirtualColumnAwareIterator to IcebergAcidUtil
